Begin the exercise for the ProgressIndicator
