+++ INSTRUCTIONS.md
@@ -1,0 +1,11 @@
+# Progress Indicator Component
+
+This is where the block comes together.
+
+You'll see the block render in the editor.
+
+## Exercise
+
+### Files
+- `js/src/edit.js`
+- `js/src/progress-indicator.exercise.js`
